# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DDBJ Search API",
    "description": "RESTful API for searching and retrieving BioProject, BioSample, SRA, and JGA entries from DDBJ.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/entries/{type}/{id}.json": {
      "get": {
        "tags": [
          "Entry Detail"
        ],
        "summary": "Get raw entry (ES document)",
        "description": "Retrieve the raw Elasticsearch document for an entry.  All fields including the full ``dbXrefs`` are returned as-is.\n\nThe response type varies by database type: BioProject, BioSample, SRA, or JGA.",
        "operationId": "get_entry_json_entries__type___id__json_get",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DbType",
              "description": "Database type."
            },
            "description": "Database type."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Entry accession identifier.",
              "title": "Id"
            },
            "description": "Entry accession identifier."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BioProject"
                    },
                    {
                      "$ref": "#/components/schemas/BioSample"
                    },
                    {
                      "$ref": "#/components/schemas/SRA"
                    },
                    {
                      "$ref": "#/components/schemas/JGA"
                    }
                  ],
                  "title": "Response Get Entry Json Entries  Type   Id  Json Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g. deep paging limit exceeded).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found (entry does not exist or invalid type).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (parameter validation error).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/entries/{type}/{id}.jsonld": {
      "get": {
        "tags": [
          "Entry Detail"
        ],
        "summary": "Get entry in JSON-LD format",
        "description": "Retrieve an entry in JSON-LD format with ``@context`` and ``@id`` fields added for RDF compatibility.\n\nContent-Type: application/ld+json.",
        "operationId": "get_entry_jsonld_entries__type___id__jsonld_get",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DbType",
              "description": "Database type."
            },
            "description": "Database type."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Entry accession identifier.",
              "title": "Id"
            },
            "description": "Entry accession identifier."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/ld+json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BioProjectEntryJsonLdResponse"
                    },
                    {
                      "$ref": "#/components/schemas/BioSampleEntryJsonLdResponse"
                    },
                    {
                      "$ref": "#/components/schemas/SraEntryJsonLdResponse"
                    },
                    {
                      "$ref": "#/components/schemas/JgaEntryJsonLdResponse"
                    }
                  ],
                  "title": "Response Get Entry Jsonld Entries  Type   Id  Jsonld Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g. deep paging limit exceeded).",
            "content": {
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found (entry does not exist or invalid type).",
            "content": {
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (parameter validation error).",
            "content": {
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error.",
            "content": {
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/entries/{type}/{id}/dbxrefs.json": {
      "get": {
        "tags": [
          "Entry Detail"
        ],
        "summary": "Get all dbXrefs",
        "description": "Retrieve all cross-references for an entry in one response.",
        "operationId": "get_dbxrefs_full_entries__type___id__dbxrefs_json_get",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DbType",
              "description": "Database type."
            },
            "description": "Database type."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Entry accession identifier.",
              "title": "Id"
            },
            "description": "Entry accession identifier."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbXrefsFullResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g. deep paging limit exceeded).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found (entry does not exist or invalid type).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (parameter validation error).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/entries/{type}/{id}": {
      "get": {
        "tags": [
          "Entry Detail"
        ],
        "summary": "Get entry detail (frontend-oriented)",
        "description": "Retrieve a single entry with dbXrefs truncated by ``dbXrefsLimit`` and ``dbXrefsCount`` added.  Use this endpoint for frontend display.  For the raw ES document, use the ``.json`` variant.\n\nThe response type varies by database type: BioProjectDetailResponse, BioSampleDetailResponse, SraDetailResponse, or JgaDetailResponse.",
        "operationId": "get_entry_detail_entries__type___id__get",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DbType",
              "description": "Database type."
            },
            "description": "Database type."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Entry accession identifier.",
              "title": "Id"
            },
            "description": "Entry accession identifier."
          },
          {
            "name": "dbXrefsLimit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 0,
              "description": "Maximum number of dbXrefs to return (0-1000). Use 0 to omit dbXrefs but still get dbXrefsCount.",
              "default": 100,
              "title": "Dbxrefslimit"
            },
            "description": "Maximum number of dbXrefs to return (0-1000). Use 0 to omit dbXrefs but still get dbXrefsCount."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BioProjectDetailResponse"
                    },
                    {
                      "$ref": "#/components/schemas/BioSampleDetailResponse"
                    },
                    {
                      "$ref": "#/components/schemas/SraDetailResponse"
                    },
                    {
                      "$ref": "#/components/schemas/JgaDetailResponse"
                    }
                  ],
                  "title": "Response Get Entry Detail Entries  Type   Id  Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g. deep paging limit exceeded).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found (entry does not exist or invalid type).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (parameter validation error).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BioProject": {
        "properties": {
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "properties": {
            "title": "Properties"
          },
          "distribution": {
            "items": {
              "$ref": "#/components/schemas/Distribution"
            },
            "type": "array",
            "title": "Distribution"
          },
          "isPartOf": {
            "type": "string",
            "const": "BioProject",
            "title": "Ispartof"
          },
          "type": {
            "type": "string",
            "const": "bioproject",
            "title": "Type"
          },
          "objectType": {
            "type": "string",
            "enum": [
              "UmbrellaBioProject",
              "BioProject"
            ],
            "title": "Objecttype"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "organism": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Organism"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "organization": {
            "items": {
              "$ref": "#/components/schemas/Organization"
            },
            "type": "array",
            "title": "Organization"
          },
          "publication": {
            "items": {
              "$ref": "#/components/schemas/Publication"
            },
            "type": "array",
            "title": "Publication"
          },
          "grant": {
            "items": {
              "$ref": "#/components/schemas/Grant"
            },
            "type": "array",
            "title": "Grant"
          },
          "externalLink": {
            "items": {
              "$ref": "#/components/schemas/ExternalLink"
            },
            "type": "array",
            "title": "Externallink"
          },
          "dbXrefs": {
            "items": {
              "$ref": "#/components/schemas/Xref"
            },
            "type": "array",
            "title": "Dbxrefs"
          },
          "sameAs": {
            "items": {
              "$ref": "#/components/schemas/Xref"
            },
            "type": "array",
            "title": "Sameas"
          },
          "status": {
            "type": "string",
            "enum": [
              "live",
              "unpublished",
              "suppressed",
              "withdrawn"
            ],
            "title": "Status"
          },
          "accessibility": {
            "type": "string",
            "enum": [
              "public-access",
              "controlled-access"
            ],
            "title": "Accessibility"
          },
          "dateCreated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datecreated"
          },
          "dateModified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datemodified"
          },
          "datePublished": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datepublished"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "properties",
          "distribution",
          "isPartOf",
          "type",
          "objectType",
          "name",
          "url",
          "organism",
          "title",
          "description",
          "organization",
          "publication",
          "grant",
          "externalLink",
          "dbXrefs",
          "sameAs",
          "status",
          "accessibility",
          "dateCreated",
          "dateModified",
          "datePublished"
        ],
        "title": "BioProject"
      },
...
      "BioProjectEntryJsonLdResponse": {
        "properties": {
          "@context": {
            "type": "string",
            "title": "@Context"
          },
          "@id": {
            "type": "string",
            "title": "@Id"
          },
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "@context",
          "@id",
          "identifier",
          "type"
        ],
        "title": "BioProjectEntryJsonLdResponse",
        "description": "BioProject entry in JSON-LD format."
      },
      "BioSample": {
        "properties": {
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "properties": {
            "title": "Properties"
          },
          "distribution": {
            "items": {
              "$ref": "#/components/schemas/Distribution"
            },
            "type": "array",
            "title": "Distribution"
          },
          "isPartOf": {
            "type": "string",
            "const": "BioSample",
            "title": "Ispartof"
          },
          "type": {
            "type": "string",
            "const": "biosample",
            "title": "Type"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "organism": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Organism"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "attributes": {
            "items": {
              "$ref": "#/components/schemas/Attribute"
            },
            "type": "array",
            "title": "Attributes"
          },
          "model": {
            "items": {
              "$ref": "#/components/schemas/Model"
            },
            "type": "array",
            "title": "Model"
          },
          "package": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Package"
              },
              {
                "type": "null"
              }
            ]
          },
          "dbXrefs": {
            "items": {
              "$ref": "#/components/schemas/Xref"
            },
            "type": "array",
            "title": "Dbxrefs"
          },
          "sameAs": {
            "items": {
              "$ref": "#/components/schemas/Xref"
            },
            "type": "array",
            "title": "Sameas"
          },
          "status": {
            "type": "string",
            "enum": [
              "live",
              "unpublished",
              "suppressed",
              "withdrawn"
            ],
            "title": "Status"
          },
          "accessibility": {
            "type": "string",
            "enum": [
              "public-access",
              "controlled-access"
            ],
            "title": "Accessibility"
          },
          "dateCreated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datecreated"
          },
          "dateModified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datemodified"
          },
          "datePublished": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datepublished"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "properties",
          "distribution",
          "isPartOf",
          "type",
          "name",
          "url",
          "organism",
          "title",
          "description",
          "attributes",
          "model",
          "package",
          "dbXrefs",
          "sameAs",
          "status",
          "accessibility",
          "dateCreated",
          "dateModified",
          "datePublished"
        ],
        "title": "BioSample"
      },
...
      "DbType": {
        "type": "string",
        "enum": [
          "bioproject",
          "biosample",
          "sra-submission",
          "sra-study",
          "sra-experiment",
          "sra-run",
          "sra-sample",
          "sra-analysis",
          "jga-study",
          "jga-dataset",
          "jga-dac",
          "jga-policy"
        ],
        "title": "DbType",
        "description": "Database types supported by the API (12 types)."
      },
...
      "ProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Problem type URI.",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Short human-readable summary."
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "HTTP status code."
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable explanation."
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance",
            "description": "Request path where the error occurred."
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp",
            "description": "Error timestamp (ISO 8601)."
          },
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid",
            "description": "Request tracking ID (same as X-Request-ID header)."
          }
        },
        "type": "object",
        "required": [
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemDetails",
        "description": "RFC 7807 Problem Details error response."
      },
...
      "SRA": {
        "properties": {
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "properties": {
            "title": "Properties"
          },
          "distribution": {
            "items": {
              "$ref": "#/components/schemas/Distribution"
            },
            "type": "array",
            "title": "Distribution"
          },
          "isPartOf": {
            "type": "string",
            "const": "sra",
            "title": "Ispartof"
          },
          "type": {
            "type": "string",
            "enum": [
              "sra-submission",
              "sra-study",
              "sra-experiment",
              "sra-run",
              "sra-sample",
              "sra-analysis"
            ],
            "title": "Type"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "organism": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Organism"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "dbXrefs": {
            "items": {
              "$ref": "#/components/schemas/Xref"
            },
            "type": "array",
            "title": "Dbxrefs"
          },
          "sameAs": {
            "items": {
              "$ref": "#/components/schemas/Xref"
            },
            "type": "array",
            "title": "Sameas"
          },
          "status": {
            "type": "string",
            "enum": [
              "live",
              "unpublished",
              "suppressed",
              "withdrawn"
            ],
            "title": "Status"
          },
          "accessibility": {
            "type": "string",
            "enum": [
              "public-access",
              "controlled-access"
            ],
            "title": "Accessibility"
          },
          "dateCreated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datecreated"
          },
          "dateModified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datemodified"
          },
          "datePublished": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datepublished"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "properties",
          "distribution",
          "isPartOf",
          "type",
          "name",
          "url",
          "organism",
          "title",
          "description",
          "dbXrefs",
          "sameAs",
          "status",
          "accessibility",
          "dateCreated",
          "dateModified",
          "datePublished"
        ],
        "title": "SRA"
      },
...
    }
  }
}